--- notification/src/main/java/org/bookwoori/notification/domain/Device.java
@@ -1,52 +1,0 @@
-package org.bookwoori.notification.domain;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.bookwoori.notification.domain.type.Platform;
-
-
-@Entity
-@Table(name = "device")
-@Getter
-@Setter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Device extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "device_id")
-    private Long id;
-
-    private Long memberId;
-
-    @Enumerated(EnumType.STRING)
-    private Platform platform;
-
-    private String token;
-
-    private boolean status;
-
-    // 생성 메서드
-    public static Device register(Long memberId, Platform platform, String token) {
-        Device device = new Device();
-        device.setMemberId(memberId);
-        device.setPlatform(platform);
-        device.setToken(token);
-        device.setStatus(true);
-        return device;
-    }
-
-    // 비즈니스 메서드
-    public void renew(Platform platform, String token) {
-        this.platform = platform;
-        this.token = token;
-        this.status = true;
-    }
-
-    public void delete() {
-        this.status = false;
-    }
-}
