+++ notification/src/main/java/org/bookwoori/notification/domain/device/dto/response/DataResponseDto.java
@@ -1,0 +1,12 @@
+package org.bookwoori.notification.domain.device.dto.response;
+
+import lombok.Builder;
+
+@Builder
+public record DataResponseDto<T>(
+        Boolean isSuccess,
+        int code,
+        String message,
+        T result
+) {
+}
