+++ notification/src/main/java/org/bookwoori/notification/domain/device/entity/Device.java
@@ -1,0 +1,52 @@
+package org.bookwoori.notification.domain.device.entity;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import org.bookwoori.notification.global.BaseTimeEntity;
+
+
+@Entity
+@Table(name = "device")
+@Getter
+@Setter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Device extends BaseTimeEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "device_id")
+    private Long id;
+
+    private Long memberId;
+
+    @Enumerated(EnumType.STRING)
+    private Platform platform;
+
+    private String token;
+
+    private boolean status;
+
+    // 생성 메서드
+    public static Device register(Long memberId, Platform platform, String token) {
+        Device device = new Device();
+        device.setMemberId(memberId);
+        device.setPlatform(platform);
+        device.setToken(token);
+        device.setStatus(true);
+        return device;
+    }
+
+    // 비즈니스 메서드
+    public void renew(Platform platform, String token) {
+        this.platform = platform;
+        this.token = token;
+        this.status = true;
+    }
+
+    public void delete() {
+        this.status = false;
+    }
+}
