+++ notification/src/main/java/org/bookwoori/notification/domain/device/entity/Platform.java
@@ -1,0 +1,12 @@
+package org.bookwoori.notification.domain.device.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum Platform {
+    WEB("WEB");
+
+    private final String description;
+}
