--- notification/src/main/java/org/bookwoori/notification/domain/dto/request/ChannelMessageRequestDto.java
@@ -1,33 +1,0 @@
-package org.bookwoori.notification.domain.dto.request;
-
-import jakarta.validation.constraints.NotNull;
-
-
-public record ChannelMessageRequestDto(
-
-        @NotNull
-        Long memberId,
-
-        @NotNull
-        String nickname,
-
-        @NotNull
-        String type,
-
-        @NotNull
-        String content,
-
-        @NotNull
-        String channelName,
-
-        @NotNull
-        Long communityId,
-
-        @NotNull
-        Long channelId,
-
-        @NotNull
-        String target
-) {
-
-}
