--- notification/src/main/java/org/bookwoori/notification/domain/dto/response/DataResponseDto.java
@@ -1,12 +1,0 @@
-package org.bookwoori.notification.domain.dto.response;
-
-import lombok.Builder;
-
-@Builder
-public record DataResponseDto<T>(
-        Boolean isSuccess,
-        int code,
-        String message,
-        T result
-) {
-}
