+++ notification/src/main/java/org/bookwoori/notification/domain/notification/dto/request/DirectMessageRequestDto.java
@@ -1,0 +1,30 @@
+package org.bookwoori.notification.domain.notification.dto.request;
+
+import jakarta.validation.constraints.NotNull;
+
+
+public record DirectMessageRequestDto(
+        @NotNull
+        Long memberId,
+
+        @NotNull
+        String nickname,
+
+        @NotNull
+        String type,
+
+        @NotNull
+        String content,
+
+        @NotNull
+        String roomName,
+
+        @NotNull
+        Long roomId,
+
+        @NotNull
+        String target
+) {
+
+
+}
