--- notification/src/main/java/org/bookwoori/notification/domain/type/Platform.java
@@ -1,12 +1,0 @@
-package org.bookwoori.notification.domain.type;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum Platform {
-    WEB("WEB");
-
-    private final String description;
-}
