+++ notification/src/main/java/org/bookwoori/notification/global/exception/ExceptionAdvice.java
@@ -4,8 +4,8 @@ package org.bookwoori.notification.global.exception;
 import com.fasterxml.jackson.databind.exc.InvalidFormatException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.bookwoori.notification.domain.dto.response.CommonResponseDto;
-import org.bookwoori.notification.domain.service.ResponseService;
+import org.bookwoori.notification.domain.notification.dto.response.CommonResponseDto;
+import org.bookwoori.notification.domain.device.service.ResponseService;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.BindException;
 import org.springframework.web.HttpRequestMethodNotSupportedException;
